+++ R/tar_store.R
@@ -1,26 +1,28 @@
-#' @title Current data store path
+#' @title Deprecated: current data store path
 #' @export
+#' @keywords internal
 #' @family utilities
-#' @description Identify the file path to the data store
-#'  of the pipeline currently running.
+#' @description Deprecated: identify the file path to the data store
+#'   of the pipeline currently running.
+#' @details `tar_store()` was deprecated on 2022-10-11 (version 0.13.5.9000).
+#'   Use [tar_path_store()] instead.
 #' @return Character, file path to the data store
 #'   of the pipeline currently running.
 #'   If called outside of the pipeline currently running,
 #'   `tar_store()` returns `tar_config_get("store")`.
 #' @examples
-#' tar_store()
+#' tar_path_store()
 #' if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
 #' tar_dir({ # tar_dir() runs code from a temporary directory.
-#' tar_script(tar_target(x, tar_store()), ask = FALSE)
+#' tar_script(tar_target(x, tar_path_store()), ask = FALSE)
 #' store <- tempfile()
 #' tar_make(store = store)
 #' tar_read(x, store = store)
 #' })
 #' }
 tar_store <- function() {
-  if_any(
-    tar_runtime$exists_store(),
-    tar_runtime$get_store(),
-    tar_config_get("store")
+  tar_warn_deprecate(
+    "tar_store() is deprecated. Use tar_path_store() instead."
   )
+  tar_path_store()
 }