+++ tests/testthat/test-tar_make.R
@@ -172,7 +172,7 @@ tar_test("custom script and store args with callr function", {
 tar_test("runtime settings are forwarded, local process", {
   skip_cran()
   tar_script({
-    writeLines(tar_store(), "store.txt")
+    writeLines(tar_path_store(), "store.txt")
     writeLines(tar_path_script(), "script.txt")
     tar_target(x, 1)
   }, script = "custom_script")
@@ -188,7 +188,7 @@ tar_test("runtime settings are forwarded, local process", {
 tar_test("runtime settings are forwarded, extrernal process", {
   skip_cran()
   tar_script({
-    writeLines(tar_store(), "store.txt")
+    writeLines(tar_path_store(), "store.txt")
     writeLines(tar_path_script(), "script.txt")
     tar_target(x, 1)
   }, script = "custom_script")